bunk commit

+++ src/main/java/com/everis/cars/control/CarService.java
@@ -5,6 +5,7 @@ import javax.ejb.Stateless;
 
 import com.everis.cars.entity.Car;
 import java.util.List;
+import com.everis.exceptions.*;
 
 @Stateless
 public class CarService {
@@ -12,28 +13,50 @@ public class CarService {
 	@PersistenceContext(unitName = "em_postgre")
 	private EntityManager em;
 	
-	public void getCars() {
+	public List<Car> getCars() {
 		 List <Car> listCars = em.createNamedQuery("Car.findAll", Car.class).getResultList();
+		 return listCars;
 	}
 	
-	public Car getCar(int id) {
+	public Car getCar(final int id) {
 		Car car = em.find(Car.class, id);
-		return car; 
+		return  car; 
 		
 	}
 	
-	public Car createCar(Car car) {
-		em.persist(car);
-		return car;
+	public Car createCar(final Car car) throws CarNotFoundException {
+		@SuppressWarnings("unchecked")
+		List <Car> listCars = em.createNamedQuery("Car.findById").setParameter("id", car.getId()).getResultList();
+		if(!listCars.contains(car)) {
+			em.persist(car);
+			return car;		
+		}else {
+			throw new CarNotFoundException(car.getId());
+		}
+		
 	}
 	
-	public Car updateCar(Car car) {
-		em.merge(car);
-		return car;
+	public Car updateCar(final Car car) throws ExistingCarException {
+		@SuppressWarnings("unchecked")
+		List<Car> rList = em.createNamedQuery("Car.findById").setParameter("id", car.getId()).getResultList();
+		if(rList.size() > 0) {
+			em.persist(car);
+			return car;
+		}else {
+			throw new ExistingCarException(car.getId()); 
+		}
 	}
 	
-	public void deleteCar(int id) {
-		em.remove(id);
+	public void deleteCar(final int id) throws NonExistingCarException {
+		Car carId = em.find(Car.class, id);
+		if(carId != null){
+			em.remove(id);
+		}else {
+			throw new NonExistingCarException(id);
+		}
+		
+		
+		
 	}
